Use CPython 3.12's test bench to check printing/parsing of Python dialect (#126)
*Issue #, if available:*

*Description of changes:*

This patch uses CPython 3.12's test bench
(https://github.com/python/cpython/blob/main/Lib/test , branch 3.12) to
test printing and parsing of Python dialect.
Concretely, the new `run_cpython_tests.sh` script
(1) Takes the .ion file output of 'python3 -m strata.gen parse'
(2) Print it to as a textual format of Strata program using `strata
print`
(3) Parse the textual format again and compare it with the ion file in
step 1, using a new `strata diff` command.

To implement this, (1) the boolean equality of the `Program` type is
added (not LawfulBEq yet - proving it for mutually inductive data types
with its constructors using List of itself was a pain....) and (2) the
new `strata diff` command is added.

There are 1052 test files in the CPython; 165 test files could not be
parsed (which is still fine, maybe?). The remaining 887 Python files
passed the roundtrip test.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.