Python AST to Boogie (#199)
Implements initial version of Python AST -> Boogie translation.
The main pain point is handling kw args. We choose a canonical ordering
for function args and use that for the Boogie representation. The code
for this is in `Strata/Languages/Python/FunctionSignatures.lean`
Ideally, we can mechanically extract this from the Boogie prelude.

Instructions for running the code are here:

StrataTest/Languages/Python/README.md

The code in `StrataTest/Internal` is empty stubs for internal code.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.