Turn Strata Python bindings into native namespace package (#276)
*Description of changes:*

This change removes `__init__.py` and refactors the imports in the
Strata Python bindings package accordingly so that it is exposed as a
[native namespace
package](https://packaging.python.org/en/latest/guides/packaging-namespace-packages/#native-namespace-packages).
Consequently, other Python packages can be added to the `strata`
namespace outside of the Strata repository.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.