PyAnalyze run multiple Z3 configs in parallel. (#288)
We need different solver configurations for some of our programs, but
with the right configuration the solver finishes quickly. This adds a
thin wrapper around Z3 that calls Z3 with various options set, each call
with a 1s timeout. Currently calls to Z3 are not raced.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.